+++ examples/fringes_transform.py
@@ -12,6 +12,8 @@ import WrightTools as wt
 from WrightTools import datasets
 
 p = datasets.PyCMDS.w2_w1_000
+
+# p = datasets.here / "PyCMDS" / "w2 w1 000.data"
 data = wt.data.from_PyCMDS(p)
 
 data.signal_mean.symmetric_root(2)  # to amplitude level
@@ -30,11 +32,11 @@ ax.set_title("as taken", fontsize=20)
 ax = plt.subplot(gs[0, 1])
 data.transform("wm", "w1")
 data.convert("wn")
-ax.pcolor(data)
+art = ax.pcolor(data)
 wt.artists.set_ax_labels(xlabel=data.wm.label, yticks=False)
 ax.grid()
 ax.set_title("transformed", fontsize=20)
 
 # colorbar
 cax = plt.subplot(gs[0, -1])
-wt.artists.plot_colorbar(cax, label="amplitude")
+fig.colorbar(art, cax, label="amplitude")