+++ CHANGELOG.md
@@ -5,12 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- `Data.ichop`, an interation-based version of `Data.chop`
+
 ### Fixed
 - wt5 explore : fixed bug where data will not load interactively if directory is not cwd
 - constants in chopped data will inherit the units of the original data
 
 ## Changed
 - artists now gets turbo colormap straight from matplotlib
+- deprecating  `artists.plot_colorbar`: instead use matplotlib's `colorbar` implementations directly
 
 ## [3.5.2]